# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Personal Music Search Engine API
  description: |
    API for a personal music library web application for uploading, organizing,
    searching, and streaming legal music files.
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: https://api.example.com/api/v1
    description: Production server
  - url: http://localhost:8080/api/v1
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /albums:
    get:
      tags: [Album]
      summary: List albums
      operationId: listAlbums
      parameters:
        - name: artist
          in: query
          schema:
            type: string
        - name: genre
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: integer
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/lastKey'
        - $ref: '#/components/parameters/sortBy'
        - $ref: '#/components/parameters/sortOrder'
      responses:
        '200':
          description: List of albums
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /albums/{id}:
    get:
      tags: [Album]
      summary: Get album with tracks
      operationId: getAlbum
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Album with tracks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumWithTracks'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /artists:
    get:
      tags: [Artist]
      summary: List artists
      operationId: listArtists
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/lastKey'
        - $ref: '#/components/parameters/sortBy'
        - $ref: '#/components/parameters/sortOrder'
      responses:
        '200':
          description: List of artists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /artists/{name}:
    get:
      tags: [Artist]
      summary: Get artist with albums and tracks
      operationId: getArtist
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Artist with albums and tracks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /playlists:
    get:
      tags: [Playlist]
      summary: List playlists
      operationId: listPlaylists
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/lastKey'
        - $ref: '#/components/parameters/sortBy'
        - $ref: '#/components/parameters/sortOrder'
      responses:
        '200':
          description: List of playlists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Playlist]
      summary: Create playlist
      operationId: createPlaylist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlaylistRequest'
      responses:
        '201':
          description: Created playlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /playlists/{id}:
    get:
      tags: [Playlist]
      summary: Get playlist with tracks
      operationId: getPlaylist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Playlist with tracks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistWithTracks'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Playlist]
      summary: Update playlist
      operationId: updatePlaylist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePlaylistRequest'
      responses:
        '200':
          description: Updated playlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Playlist]
      summary: Delete playlist
      operationId: deletePlaylist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Playlist deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /playlists/{id}/tracks:
    post:
      tags: [Playlist]
      summary: Add tracks to playlist
      operationId: addTracksToPlaylist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddTracksToPlaylistRequest'
      responses:
        '200':
          description: Updated playlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Playlist]
      summary: Remove tracks from playlist
      operationId: removeTracksFromPlaylist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemoveTracksFromPlaylistRequest'
      responses:
        '200':
          description: Updated playlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /playlists/{id}/tracks/reorder:
    put:
      tags: [Playlist]
      summary: Reorder tracks in playlist
      operationId: reorderPlaylistTracks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderPlaylistTracksRequest'
      responses:
        '200':
          description: Updated playlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /search:
    get:
      tags: [Search]
      summary: Simple text search
      operationId: searchSimple
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 500
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Search]
      summary: Advanced search with filters
      operationId: searchAdvanced
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /search/suggest:
    get:
      tags: [Search]
      summary: Get autocomplete suggestions
      operationId: getSearchSuggestions
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
      responses:
        '200':
          description: Autocomplete suggestions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutocompleteResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    lastKey:
      name: lastKey
      in: query
      schema:
        type: string
    sortBy:
      name: sortBy
      in: query
      schema:
        type: string
    sortOrder:
      name: sortOrder
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
    Pagination:
      type: object
      properties:
        limit:
          type: integer
        lastKey:
          type: string
        nextKey:
          type: string
        totalEstimate:
          type: integer
...
    TrackResponse:
      type: object
      required: [id, title, artist, duration, durationStr, format, fileSize, fileSizeStr, playCount, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        artist:
          type: string
        albumArtist:
          type: string
        album:
          type: string
        albumId:
          type: string
          format: uuid
        genre:
          type: string
        year:
          type: integer
        trackNumber:
          type: integer
        discNumber:
          type: integer
        duration:
          type: integer
        durationStr:
          type: string
        format:
          type: string
        fileSize:
          type: integer
          format: int64
        fileSizeStr:
          type: string
        coverArtUrl:
          type: string
          format: uri
        playCount:
          type: integer
        lastPlayed:
          type: string
          format: date-time
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AlbumResponse:
      type: object
      required: [id, title, artist, trackCount, totalDuration, durationStr, discCount, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        artist:
          type: string
        albumArtist:
          type: string
        genre:
          type: string
        year:
          type: integer
        coverArtUrl:
          type: string
          format: uri
        trackCount:
          type: integer
        totalDuration:
          type: integer
        durationStr:
          type: string
        discCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AlbumWithTracks:
      type: object
      required: [album, tracks]
      properties:
        album:
          $ref: '#/components/schemas/AlbumResponse'
        tracks:
          type: array
          items:
            $ref: '#/components/schemas/TrackResponse'
    AlbumListResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AlbumResponse'
        pagination:
          $ref: '#/components/schemas/Pagination'
    ArtistSummary:
      type: object
      required: [name, trackCount, albumCount]
      properties:
        name:
          type: string
        trackCount:
          type: integer
        albumCount:
          type: integer
        coverArtUrl:
          type: string
          format: uri
    ArtistDetail:
      type: object
      required: [name, albums, tracks, albumCount, trackCount]
      properties:
        name:
          type: string
        albums:
          type: array
          items:
            $ref: '#/components/schemas/AlbumResponse'
        tracks:
          type: array
          items:
            $ref: '#/components/schemas/TrackResponse'
        albumCount:
          type: integer
        trackCount:
          type: integer
    ArtistListResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ArtistSummary'
        pagination:
          $ref: '#/components/schemas/Pagination'
    PlaylistResponse:
      type: object
      required: [id, name, trackCount, totalDuration, durationStr, isPublic, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        coverArtUrl:
          type: string
          format: uri
        trackCount:
          type: integer
        totalDuration:
          type: integer
        durationStr:
          type: string
        isPublic:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PlaylistWithTracks:
      type: object
      required: [playlist, tracks]
      properties:
        playlist:
          $ref: '#/components/schemas/PlaylistResponse'
        tracks:
          type: array
          items:
            $ref: '#/components/schemas/TrackResponse'
    CreatePlaylistRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        isPublic:
          type: boolean
          default: false
    UpdatePlaylistRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        isPublic:
          type: boolean
    AddTracksToPlaylistRequest:
      type: object
      required: [trackIds]
      properties:
        trackIds:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: string
            format: uuid
        position:
          type: integer
          minimum: 0
    RemoveTracksFromPlaylistRequest:
      type: object
      required: [trackIds]
      properties:
        trackIds:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: string
            format: uuid
    ReorderPlaylistTracksRequest:
      type: object
      required: [trackId, newPosition]
      properties:
        trackId:
          type: string
          format: uuid
        newPosition:
          type: integer
          minimum: 0
    PlaylistListResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PlaylistResponse'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    SearchRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 500
        filters:
          type: object
          properties:
            artists:
              type: array
              items:
                type: string
            albums:
              type: array
              items:
                type: string
            genres:
              type: array
              items:
                type: string
            tags:
              type: array
              items:
                type: string
            years:
              type: array
              items:
                type: integer
            formats:
              type: array
              items:
                type: string
        sort:
          type: object
          properties:
            field:
              type: string
              enum: [relevance, title, artist, album, year, playCount, createdAt]
            order:
              type: string
              enum: [asc, desc]
        limit:
          type: integer
          minimum: 1
          maximum: 100
        offset:
          type: integer
          minimum: 0
    SearchResponse:
      type: object
      required: [query, totalResults, tracks, limit, offset]
      properties:
        query:
          type: string
        totalResults:
          type: integer
        tracks:
          type: array
          items:
            $ref: '#/components/schemas/TrackResponse'
        albums:
          type: array
          items:
            $ref: '#/components/schemas/AlbumResponse'
        artists:
          type: array
          items:
            $ref: '#/components/schemas/ArtistSummary'
        facets:
          type: object
          properties:
            artists:
              type: array
              items:
                $ref: '#/components/schemas/FacetItem'
            albums:
              type: array
              items:
                $ref: '#/components/schemas/FacetItem'
            genres:
              type: array
              items:
                $ref: '#/components/schemas/FacetItem'
            years:
              type: array
              items:
                $ref: '#/components/schemas/FacetItem'
            tags:
              type: array
              items:
                $ref: '#/components/schemas/FacetItem'
        limit:
          type: integer
        offset:
          type: integer
    FacetItem:
      type: object
      required: [value, count]
      properties:
        value:
          type: string
        count:
          type: integer
    AutocompleteResponse:
      type: object
      required: [query, suggestions]
      properties:
        query:
          type: string
        suggestions:
          type: array
          items:
            type: object
            required: [text, type]
            properties:
              text:
                type: string
              type:
                type: string
                enum: [track, artist, album, tag]
              id:
                type: string
              imageUrl:
                type: string
                format: uri
...